Moved overplot functinality to interface file
