+++ mslice/views/slice_plotter.py
@@ -1,9 +1,7 @@
 from __future__ import (absolute_import, division, print_function)
 
-from mslice.models.labels import get_recoil_label
 import mslice.plotting.pyplot as plt
 
-OVERPLOT_COLORS = {1: 'b', 2: 'g', 4: 'r', 'Aluminium': 'g', 'Copper': 'm', 'Niobium': 'y', 'Tantalum': 'b'}
 PICKER_TOL_PTS = 5
 
 
@@ -52,23 +50,3 @@ def _show_plot(slice_cache, workspace):
 def set_colorbar_label(label):
     plt.gcf().get_axes()[1].set_ylabel(label, rotation=270, labelpad=20)
 
-
-def remove_line(line):
-    plt.gca().lines.remove(line)
-
-
-def plot_overplot_line(x, y, key, recoil, cache):
-    color = OVERPLOT_COLORS[key] if key in OVERPLOT_COLORS else 'c'
-    if recoil:
-        return overplot_line(x, y, color, get_recoil_label(key), cache.rotated)
-    else:
-        return overplot_line(x, y, color, key, cache.rotated)
-
-
-def overplot_line(x, y, color, label, rotated):
-    if rotated:
-        return plt.gca().plot(y, x, color=color, label=label, alpha=.7,
-                              picker=PICKER_TOL_PTS)[0]
-    else:
-        return plt.gca().plot(x, y, color=color, label=label, alpha=.7,
-                              picker=PICKER_TOL_PTS)[0]
